Created new xref route and moved xref_collection_unit getter into there, moved oneDay constant to constants file
