+++ src/api/collectionUnit/collectionUnit.service.ts
@@ -51,22 +51,6 @@ const getCollectionUnitsByCritterId = async (
   return collectionUnits;
 };
 
-const getCollectionUnitsFromCategory = async (category_name: string, taxon_name_common: string | undefined, taxon_name_latin: string | undefined) => {
-  const taxon_categories = await prisma.xref_taxon_collection_category.findFirstOrThrow({
-    where: {
-      lk_taxon:  { taxon_name_common: taxon_name_common, taxon_name_latin: taxon_name_latin },
-      lk_collection_category: { category_name: {equals: category_name, mode: 'insensitive' } }
-    }
-  });
-  const category_id = taxon_categories?.collection_category_id;
-  const collectionUnits = await prisma.xref_collection_unit.findMany({
-    where: {
-      collection_category_id: category_id
-    }
-  });
-  return collectionUnits;
-}
-
 /**
  * * Updates an existing critter collection unit in the database
  * @param {string} critter_collection_unit_id
@@ -120,7 +104,6 @@ const deleteCollectionUnit = async (
 export {
   getAllCollectionUnits,
   getCollectionUnitById,
-  getCollectionUnitsFromCategory,
   getCollectionUnitsByCritterId,
   updateCollectionUnit,
   createCollectionUnit,
